refactor: path name to tracker and typedef adjust
